requirements.txt
# LiveKit Agent Framework
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-elevenlabs>=0.6.0

# Google Sheets Integration
gspread>=5.11.0
google-auth>=2.23.0

# Environment & Config
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0

# Async Support
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
